relocate voting utils
