+++ app/src/lib/utils/voting.ts
